Clear Cache + Include other files
Set up the cache to be cleared if going from the file select screen to the title screen. This is needed to make sure that it gets cleared upon making another change to the settings without actually loading a file.

Also added in the other files from ttyd-tools, some of which are needed to build this project.