Adjust Assembly Codes
Modified the assembly codes to be as short as possible, as they should be easily modifiable from the C++ code.

I purposely skipped the enemy held item assembly codes, as those codes need to be redone anyway to fix a bug.

Also shifted the game mode text up a bit.