Various changes
Allowed the timer to continue even on the file select screen and when loading a file, unless the file is a new file.

The game is now paused and control is taken away from the player when the time is up.

Allows the Time's Up text to appear in battles.

All partners are now given to the player upon entering a loading zone.

Fixed an issue with the marioKeyOn function, in which the key could become negative when being used with the timer code.

Disabled the following in order to prevent crashes:
1. Blue switches in the Grotto room leading to the exit in the wall.
2. Blue switch in the water switch room in the Great Tree.
3. Pipe in the fake Garnet Star room

Prevented enemies from being able to hold the Debug Badge.

Created a hotfix for item drops not working, until the reason for them bugging out is found.