More fixes + more adjustments
Added more stuff to try to fix the enemy item drops.

The score, sequence, loading zone, and map are now displayed while in battles.

Fixed the game crashing at the end of the Mowz cutscene in Hooktail Castle.

Fixed the game softlocking if a partner is not out during the Koopie Koo cutscene.

Moved Jabble under the map to prevent the game from crashing when talking to him a wrong partner out.