More Loading Zone Randomizer changes
It is now impossible to warp directly to the credits if not using the challenge mode.

The Shadow Queen can now be fully fought without needing to set the sequence. Most of the cutscenes regarding fighting her are also skipped.

The Atomic Boo is now counted as a boss that you can get points from.

The Shadow Queen and Bonetail now give 20 points instead than 10.