Major Loading Zone Randomizer changes
Gloomtail and the Shadow Sirens from the Palace of Shadow are now counted as bosses.

The cutscene of Crump flying off into space after beating Magnus 2.0 is now skipped.

The sequence is now automatically adjusted for bosses upon entering the rooms for them. However, if the challenge mode is being used and 20 minutes have passed, then these rooms will not be possible to visit if the current sequence allows for them to be fought.

Only the first character of the file name has to be the star in order to prevent warping to the credits.