Additional LZ Randomizer Check
Prevent warping to the room with the Atomic Boo if using the challenge mode, 20 minutes have not passed, and if the player currently has a hammer upgrade.

This is to ensure that the Atomic Boo cannot actually be fought until 20 minutes have passed, but the player can still access the room if they do not have a hammer upgrade.