Various changes
Created a function to check if the challenge mode is being used and if 20 minutes have passed, as these checks are used a lot.

Added the Strange Sack back into the item pool, as it should still be included if the Loading Zone randomizer is not in use.

Adjusted the points given for level ups from 3 to 2.

Prevented the Magical Map or crystal stars from spawning within the first 5 minutes, assuming that the challenge mode is in use.

Adjusted the SP needed for Sweet Feast from 5 to 3.

Adjusted the SP needed for Showstopper from 2 to 4.

Prevented the escape timer from appearing if using the Loading Zone randomizer.

Made adjustments so that the Paper curse can be obtained much more easily.