Major changes
Fixed a bug with Magnus 1 that would allow the player to get a Game Over and then be able to get 10 points for defeating him.

Prevented the Debug Badge from appearing if the challenge mode is in use.

Manually set the Sequence when entering the first screens of Twilight Town and the Palace of Shadow to skip their cutscenes.

Removed unnecessary checks for the Hooktail Castle GSWFs.

Prevented being able to reload the screen when receiving an item. This is to prevent the player from opening a chest and then reloading the room to be able to open it again.

Used the proper function to check for Hammer upgrades.

Fixed a bug with reloading the room after opening a Curse chest, which would allow being able to talk to the chest again.

Added functions to handle the commonly-used functions for drawing text.