Loading Zone randomixer changes
Prevented the following cutscenes from playing:
1. Fahr Outpost intro area (with the pipe)
2. Riverside Station with Doopliss
3. X-Naut Fortress intro
4. Petal Meadows intro
5. Petal Meadows Hooktail flying

Changed the loading zone of the room with the Ultra Boots to only be changed if the player has not opened the chest yet, and if the Sequence is not above the proper point for the chest to spawn.

Changed the partner brought out for Bonetail to Goombella.

Goombella is now brought out if no partner is currently out when warping to the Shadow Queen cutscene.

5 points are now lost per Game Over.

The Score, Sequence, Loading Zone, and Map are now displayed on the Game Over screen.