LZ Rando changes
Changed the title screen text to use the rainbow effect.

Skipped the intro cutscene outside of the Creepy Steeple.

Skipped the cutscene of the Shadow Sirens stealing the fake Garnet Star.

The Cog is now already placed, and the crane can be used.

Skipped both cutscenes of signing up to be a fighter.