Major fixes and changes
Fixed a major bug that would occur when using First Attack or Bump Attack.

Removed Item Hog from the possible random items.

The Sequence is now changed if the player Game Overs on the Shadow Queen.

It is now impossible to reload the room to refight bosses.

Prevented the player from changing the loading zone during Magnus 2 and the Shadow Queen boss fights, as they need specific loading zones to function properly.

When a file is loaded when using the Loading Zone randomizer, Mario's coordinates are reset to 0, 0, 0.

Collecting the Crystal Star reward at the end of each chapter no longer gives the special move, nor does it adjust the max SP.