Major changes
Simplified the functions that handle drawing text.

Adjusted the versions text on the title screen to be placed on top of a window.

Added the first room of the moon to the Loading Zone randomizer.

Prevented the player from being able to use the Reload Room code inside of the Shadow Queen's room while the Sequence is set for her.

Included the overall version of the file on the title screen.

Having 999 coins now gives you an additional point.

Fixed the crashes revolving around Doopliss 2: The game now forces Vivian to be out for both cutscenes.

Goombella will no longer be brought out at the start of the Doopliss 2 fight.

Skipped the Crane Game tutorial.

Skipped the cutscene with Four-Eyes after Bobbery is taken by the Embers.

Skipped Vivian talking to Mario in Twilight Trail.

Skipped the Trouble Center tutorial.

Followers are now randomized (you don't start with one; you still have to be given one normally.)

It is no longer possible to continuously reload the room with the Blue Key to get an unlimited amount of them.

Many cutscenes that were previously skipped are no longer skipped in the standard loading zone randomizer. They are still skipped in the challenge mode.

Many maps have been removed from the challenge mode. The list is as follows:
1. aji_09 - X-Naut Fortress - Changing Room
2. aji_10 - X-Naut Fortress - Control Room/X-Yux
3. aji_16 - X-Naut Fortress - Locker Room
4. aji_17 - X-Naut Fortress - TEC's Room
5. aji_19 - X-Naut Fortress - Conveyor Belt Entrance
6. dou_01 - Pirate's Grotto - 2 Springs
7. dou_08 - Pirate's Grotto - Outside Sunken Ship
8. dou_12 - Pirate's Grotto - Bridge with Bill Blaster
9. eki_01 - Riverside Station - Tube Mode Blue Switch
10. end_00 - Credits
11. hei_07 - Shhwonk Fortress - Entrance/Quiz Room
12. las_00 - Palace of Shadow - Entrance
13. las_05 - Palace of Shadow - Dark Bones/Locked Door
14. las_08 - Palace of Shadow - Endless Room Puzzle
15. las_27 - Palace of Shadow - Save Block/Recovery Block/Before Grodus
16. las_28 - Palace of Shadow - Grodus's Room
17. mri_08 - The Great Tree - Punies vs 10 Jabbies
18. mri_11 - The Great Tree - Punies vs 100 Jabbies
19. mri_14 - The Great Tree - Lowest Chamber
20. muj_11 - Pirate's Grotto - Cortez's Ship Entrance
21. pik_02 - Poshley Sanctum - Fake Garnet Star Room
22. rsh_00_a - Excess Express - Right Engineer's Car (Day)
23. rsh_00_b - Excess Express - Right Engineer's Car (Dusk)
24. rsh_01_a - Excess Express - Cabins #1-2 (Day)
25. rsh_01_b - Excess Express - Cabins #1-2 (Dusk)
26. rsh_01_c - Excess Express - Cabins #1-2 (Night)
27. rsh_02_a - Excess Express - Cabins #3-5 (Day)
28. rsh_02_b - Excess Express - Cabins #3-5 (Dusk)
29. rsh_03_a - Excess Express - Dining Car (Day)
30. rsh_03_b - Excess Express - Dining Car (Dusk)
31. rsh_04_a - Excess Express - Cabins #6-8 (Day)
32. rsh_04_b - Excess Express - Cabins #6-8 (Dusk)
33. rsh_07_a - Excess Express - Left Engineer's Car (Day)
34. rsh_07_b - Excess Express - Left Engineer's Car (Dusk)
35. rsh_07_c - Excess Express - Left Engineer's Car (Night)
36. rsh_08_a - Excess Express - Right Freight Car
37. tou_10 - Glitzville - Minor-League Locker Room
38. win_04 - Boggly Woods - Flurrie's House - Entrance
39. win_06 - Boggly Woods - Pipe Entrance

Made a separate array for the challenge mode maps.