More Changes
Changed the getRandomWarp function to not recheck which array to use when rechoosing a room.

Properly skipped the Creepy Steeple intro cutscene.

The entrances in tik_03 that Punio would normally have to reveal are now revealed by default.

In gon_05, the loading zone used for entering the room is now randomized between the one needed for the Star Piece and the one needed for lowering the block to get the key.

Removed more rooms:
1. aji_01 - X-Naut Fortress - Elevator Corridor
2. aji_18 - X-Naut Fortress - 3 Card Keys Room
3. dou_03 - Pirate's Grotto - Waterfall/Use Gate Handle
4. dou_11 - Pirate's Grotto - Blowupable Wall leading to the Exit
5. dou_13 - Pirate's Grotto - Corridor with 2 Bill Blasters
6. hei_09 - Shhwonk Fortress - West Room leading to Moon Stone Room
7. las_10 - Riddle Tower - Entrance
8. las_21 - Palace of Shadow - Corridor #2/Yellow,Green,Purple,Red Blocks
9. las_23 - Palace of Shadow - Corridor #3/Spring Jump Poles
10. moo_05 - Moon - Empty Rocks
11. mri_18 - The Great Tree - Pipe to Water Switch Room
12. tik_12 - Rogueport Sewers - Merlee's House