LZ Rando Changes
Added Merlee's room back into the challenge mode.

Prevented the score from updating when time is up.

Prevented warping to the Pit unless 10 minutes have passed.

When warping to Glitzville, the Sequence will now be adjusted for the player to be able to get the Super Hammer if the Sequence is below that point and they don't have it already.

Partners now start with the correct HP.

The proper loading zones are now used when spawning in the tile rooms in X-Naut Fortress.

The west loading zone is now used when spawning in las_03.