More LZ Rando Changes
When warping to Glitzville, the Sequence is now adjusted to allow the player to get the Super Hammer if they don't have it, regardless of what the Sequence is currently at.

Added the Shadow Sirens from Boggly Woods to the bosses that give points.

When spawning in the spike room in the sewers, the loading zone is now set to south.

The rooms with Blooper and Gloomtail can now be accessed without 20 minutes needing to have passed. The player will still not be able to fight them until 20 minutes have passed, however.