Major changes
Removed the credits warp from the loading zone randomizer without using the challenge mode.

Prevented being able to fight bosses if 20 minutes have not passed, but the player has the exact sequence needed for a boss.

Changed how Doopliss 1 and 2 are picked: one is now guarenteed to be picked if 20 minutes have passed. If the Sequence is under or at 199, then one is picked at random. If the Sequence is past 199, then Doopliss 2 is chosen.

Changed the battle trunks so that 10 of them give 1 point.

Skipped the cutscene of Goombella explaining her tattles on the bridge screen in Petal Meadows.

Fixed the game softlocking when trying to reload the screen on the same frame as entering the battle.

Set fixed loading zones used with the following maps:
1. bom_03 - Fahr Outpost - Field West Side
2. bom_04 - Fahr Outpost - Field East Side
3. dou_02 - Pirate's Grotto - Floor Spikes
4. eki_04 - Riverside Station - Maze/P-Up, D-Down
5. gon_08 - Hooktail Castle - Green Block/Life Shroom
6. gra_01 - Twilight Trail - 3 Goombas/Super Shroom Item Box
7. gra_02 - Twilight Trail - Fallen Tree
8. gra 03 - Twilight Trail - 2 Dayzees/Earthquake Item Box
9. hei_00 - Petal Meadows - Pipe To Hooktail Castle
10. hei_02 - Petal Meadows - Pow Block in Item Box
11. hei_04 - Petal Meadows - Pow Block in Bush
12. hei_08 - Shhwonk Fortress - Moon Stone Room
13. hei_12 - Shhwonk Fortress - Sun Stone Room
14. hei_13 - Petal Meadows - Close Call Item Box
15. jin_07 - Creepy Steeple - 200 Boos Chest
16. jin_10 - Creepy Steeple - Buzzy Beetles/Tornado Jump
17. las_04 - Palace of Shadow - Fire Bar Room/Boo Sheet
18. las_19 - Palace of Shadow - Corridor #1/Wizzerd/Blowable Wall (only when the Sequence is under 390)
19. moo_02 - Moon - Pipe To X-Naut Fortress
20. tik_07 - Rogueport Sewers - West Side Entrance
21. tik_15 - Rogueport Sewers - Pipe to Staircase Switch/Black Key
22. tou_04 - Glitzville - Backstage Corridor