Major Changes
Changed the loading zone for gra_02 to only be set if the Sequence is below the value set by blowing away the tree with Flurrie.

Fixed the loading zone for moo_02 not being set properly.

Removed moo_00 from the challenge mode.

The Sequence is now automatically set to allow the player to get the Ultra Boots if it is currently past 318.

jin_00 has now been adjusted so that the player will be able to get the Steeple Key. However, it wont be possible to actually use the key.

usu_01 has now been adjusted so that the player can get the Tube Mode curse without needing to lower the Sequence first, assuming that they have the black key.

Made adjustments to the functions that clear cache.

Changed some functions to properly use SeqIndex.

Changed the loading zones to be fixed for the following maps:
1. aji_19 - X-Naut Fortress - Conveyor Belt Entrance
2. eki_02 - Riverside Station - Large Clock/Station Key
3. gon_02 - Hooktail Castle - Power Bounce Item Box
4. las_21 - Palace of Shadow - Corridor #2/Yellow,Green,Purple,Red Blocks (only when the Sequence is before 390)
5. las_23 - Palace of Shadow - Corridor #3/Spring Jump Poles (only when the Sequence is before 390)
6. las_30 - Palace of Shadow - Long Stairway to Shadow Queen
7. rsh_06_a - Excess Express - Train's Roof (only when the Sequence is after 332)
8. tik_20 - Rogueport Sewers - 3 Shine Sprite Room