Major changes
Prevented warping into the Bowser cutscene in pik_01 if using the challenge mode.

Added Mega Jump, Mega Smash, and Mega Quake to the item pool.

Changed the loading zone for gon_11 to be set to west if not fighting Hooktail.

Randomized the Power Rush and Power Rush P icons for the US and EU versions when using the Loading Zone randomizer. They are randomized upon starting a new file. The icons used are from the other languages of the game.