Item Rando Changes
Fixed a bug that would cause a crash if defeating Blooper and one/both of his tentacles at the same time via Quake Hammer/Mega Quake.

Added in a badge (decimal 336, hex 0x150) that when equipped, automatically fills the run meter. The badge costs 1 BP to equip.

Removed an unnecessary function that was used for checking if the Art Attack hitboxes badge was equipped or not.