Major LZ Rando changes
The Final Score is now displayed when time is up. The original Score value, as well as the sources for them, will still update as normal.

The timer is now reset to 1 hour upon selecting to continue playing after the initial hour is up. The Final Score will remain untouched, however.

Jump Storage is now cleared upon starting a new file.

The lottery is now reset upon starting a new file (as this can apparently not be reset in rare instances).

The cutscene of the bridge being shown at Riverside after talking to the Toad is now skipped.

Having to talk to the curse chests before being able to use the keys on them is now skipped. The Boat Mode curse chest is an exception to this, as talking to it is required to spawn the ember.

It is now possible to reobtain the followers when entering the pipe in gor_02 or entering the red door in mri_00.

Set the loading zone for mri_12 to be set to east (top one).

It is no longer possible to warp to Gloomtail's room if the Sequence is exactly 400. This is because the Shadow Queen cutscene will occur if this room is entered with the Sequence at 400.