Major Changes
Set up the Help menu to use an array for the pages instead of a switch. Also moved the array into arrays.cpp.

Used keyGetButtonTrg in the proper places, and removed the DenyInput bool, as it is obsolete.

Skipped the cutscene of the Dried Shroom with Petuni when using the challenge mode (the one that normally plays after leaving the room with the Blue Key).