Major Changes
Changed the points for several bosses:
1. Grodus and Bowser & Kammy are now worth 8 points each.
2. Shadow Queen is now worth 25 points,
3. Bonetail is now worth 30 points.

Added in a separate function to call windowDispGX_Waku_col, which takes ints instead of floats. This is so that the function will convert the ints to floats manually, rather than the floats being stored globally.

Completely rewrote the drawStringMultiline function in drawstring.cpp.

Removed an unnecessary check in the setUpNewFile function.