Major Changes
Added in a menu that displays all of the player's previous final scores. It is toggled by pressing L + Y while at the File Select menu. The scores are listed in the order that the runs are completed, so the first one listed is the score from the first run, etc. If the player got a Game Over and started a new file, then the entry for that run will be Reset instead of an actual score. In the event that the total amount of scores exceed the amount that can be stored on one page (currently holds up to 64 scores), then D-Pad Left-Right can be used to navigate to the other pages. If the game is reset, then this list will also be reset, as the game's memory is cleared upon a reset. All of this info is also listed in the Help menu.

When going to fight Doopliss 2, Vivian will now only be forced out when the cutscene with Doopliss is about to start.

Added text about how Doopliss 1/2 is chosen for the Loading Zone randomizer to the Help menu.

Made the window for the Help menu slightly darker.

Removed several global strings to use stack VLA instead.