Major Changes - v2.0 Release
All loading zones are now randomized to one of the valid ones in the current room.

When going to a new room, LZ will temporarily say 'Pending' until a valid loading zone is chosen.

Removed the following codes from the standard Loading Zone randomizer (they are still included in the Challenge Mode):
1. Reloading the room
2. Manually changing the loading zone
3. Walking on water

When using the Loading Zone randomizer but not the Challenge Mode, Mario will trasnform into the Boat when coming out of a loading zone that requires it.

Paper Mode and Tube Mode will be enabled for the following rooms:
1. aji_13 - X-Naut Fortress - Factory
2. jin_05 - Creepy Steeple - Storage/Cookbook
3. jin_07 - Creepy Steeple - 200 Boos Chest

Boat Mode will be enabled for the following rooms:
1. dou_05 - Pirate's Grotto - Save Block/Locked Door
2. tik_20 - Rogueport Sewers - 3 Shine Sprite Room

Plane Mode will be enabled for the following rooms:
1. aji_13 - X-Naut Fortress - Factory
2. dou_11 - Pirate's Grotto - Blowupable Wall leading to the Exit