Major Changes
Skipped the cutscene at the bottom floor of mri_17.

Changed the fadeEntry parameter from uint8_t* to uint8_t[4].

For closing the curtain upon loading a file and for checking if the System Flag should be reset, GSW(4) is now checked instead of the Sequence.

Set up local pointers to strings to use to point to NextMap, NextBero, NextArea, and DisplayBuffer. This ends up saving over 1000 bytes.