Major Changes
Added proper names and descriptions for all of the unused items/badges. Also fixed the type sorting for all of them as well.

Fixed the game crashing when starting the Crump 2 cutscene with no partner out.

When using the challenge mode, it is no longer possible to warp to the Major League room while the Sequence is under 130. This is to avoid the cutscene with Grubba.

Fixed issues with the Boat Mode curse chest that would be caused by leaving the room after talking to the chest and then coming back when the Sequence is not 246. The Sequence will now be automatically set to 246 to get around these issues.

Partners and followers are no longer removed when in a Chapter 4 map when the Sequence is between 201-205.

While using the Loading Zone randomizer, Paper and Tube Modes are now enabled in the Creepy Steeple room with the Save Block and Shine Sprite.

Prevented the black Mario head from being used in screen transitions, as it may cause crashes with certain maps.

Added a code to the Challenge Mode that allows you to lock the current LZ, so that it will not be randomized when reloading the room. This is activated by pressing L+Y.

Added a code to the Loading Zone randomizer that allows you to toggle the on-screen text to be transparent. This is activated by pressing Z+D-Pad Up.

Prevented being able to use the codes to manually change the loading zones or lock it while a file is not loaded.

Made some additional edits to the Help menu.