Bug Fix + Backtrace Screen Edits
When going to fight Magnus 1, the loading zone is now locked to e_bero. his should prevent the game from crashing, which could occur if the Magnus cutscene were to start immediately upon entering the room.

Made the following edits to the US and EU versions of the backtrace screen:
1. The text will reloop after several seconds, instead of running only once.
2. The text is now smaller. This should prevent it from going offscreen.