Major Changes
Fixed the Red Bones being able to spawn before 20 minutes have passed.

Prevented the cutscene from playing in the Thousand-Year Door room if the Sequence is 15. The Sequence will now be set to 16 in this instance.

Added a timer to the title screen that displays how much time is left before gameplay resumes after a game over.