Major Changes
The timer on the Title Screen is now only shortened when the player gets a Game Over and is using the Loading Zone randomizer.

Fixed being able to reload the room to be able to fight the Atomic Boo before 20 minutes have passed.

Mario will now be forced into Boat Mode when entering tik_20 when the challenge mode is being used.

Obtaining the Super Boots/Hammer after the Ultra Boots/Hammer will no longer cause you to use the Super upgrade in battles.