Major Changes
Rewrote some of the functions that are used to draw text and windows.

Made separate global functions for comparing and writing strings.

Modified the code that replaces parts of the REL files with code that writes nops instead.

All of the chests in the Pit can now be opened more than once when using the Loading Zone randomizer.

After going through a pipe, when warping to either the Shadow Queen or the Crump fight in Chapter 5, the previous partner you had out will now be brought out instead of Goombella.

Followers now count for points even when their sprite is currently not loaded.

It is no longer possible to warp to las_19 and las_25 if the Sequence is under 390 while using the challenge mode.

Partners and Followers are no longer removed when warping to a gra map when the Sequence is set to just after beating Doopliss 1.

Fixed some of the README links not working.

Fixed a misnamed item in the item list.

Prevented being able to get the Elevator Key in Riverside more than once.

Removed eki_02 from the challenge mode.

Allowed the Champ's Belt to be obtained once for every time Rawk Hawk is defeated. The player must warp to the Champ's room to get the next belt. Reloading the room will not give the player the belt.

Paper Mode and Tube Mode are now enabled when in east Rogueport when using the Loading Zone randomizer.

The following bosses have had parts of their cutscenes skipped when using the challenge mode:
1. Hooktail
2. Magnus 1.0
3. Rawk Hawk
4. Grubba
5. Doopliss 2
6. Cortez
7. Crump
8. Smorg
9. Magnus 2.0
10. Bowser & Kammy

Also, if one of the listed bosses would result in the player getting an item, then they still be given an item. If the item that would normally be given has already spawned, then the player will get that item. If not, then the player will get a random item.