Challenge Mode Changes
Fixed being able to warp to gon_07 after getting the black key if using the challenge mode.

Fixed being able to warp to aji_02 if the Elevator Key has already been obtained when using the challenge mode. Now it will work as intended: getting both of the items in that room will prevent being able to warp to that room again.

Changed Grodus and Bowser & Kammy to be worth 9 points each.

Changed the Shadow Queen to be worth 30 points.

Removed gon_09 and pik_01 from the challenge mode.

When using the challenge mode, gon_08 will no longer be accessible once the item in it has been obtained.