Minor Fixes
Fixed the intro cutscene playing in jin_00 if the player had entered after having gotten a hammer upgrade, they have gotten the key, and 20 minutes have passed.

Fixed the intro cutscene playing in mri_00 if the sequence was exactly 73. The sequence will now be set to 74 in this instance.

Prevented being able to get the letter p from jin_08 more than once when using the challenge mode.